requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
distlib==0.3.9
Django==5.2.3
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-filter==25.1
django-formtools==2.5.1
django-phonenumber-field==8.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
easy-thumbnails==2.10
filelock==3.16.1
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
oauthlib==3.2.2
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==76.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sorl-thumbnail==12.11.0
sqlparse==0.5.1
stripe==12.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.30.0
Werkzeug==3.1.3
